requirements.txt.txt
streamlit
pytesseract
Pillow
opencv-python-headless
pandas
numpy